Redesign search interface to be usable by plugins
