+++ shesmu-pluginapi/src/main/java/ca/on/oicr/gsi/shesmu/plugin/filter/FilterCheckedAgo.java
@@ -1,0 +1,12 @@
+package ca.on.oicr.gsi.shesmu.plugin.filter;
+
+import java.time.Instant;
+import java.util.Optional;
+
+public class FilterCheckedAgo extends AgoFilterJson {
+  @Override
+  public <F> F convert(
+      Optional<Instant> start, Optional<Instant> end, FilterBuilder<F> filterBuilder) {
+    return filterBuilder.checked(start, end);
+  }
+}
