+++ shesmu-pluginapi/src/main/java/ca/on/oicr/gsi/shesmu/plugin/filter/FilterText.java
@@ -1,0 +1,33 @@
+package ca.on.oicr.gsi.shesmu.plugin.filter;
+
+import java.util.regex.Pattern;
+
+public class FilterText extends FilterJson {
+  private boolean matchCase;
+  private String text;
+
+  @Override
+  public <F> F convert(FilterBuilder<F> filterBuilder) {
+    return maybeNegate(
+        filterBuilder.textSearch(
+            Pattern.compile(
+                "^.*" + Pattern.quote(text) + ".*$", matchCase ? 0 : Pattern.CASE_INSENSITIVE)),
+        filterBuilder);
+  }
+
+  public String getText() {
+    return text;
+  }
+
+  public boolean isMatchCase() {
+    return matchCase;
+  }
+
+  public void setMatchCase(boolean matchCase) {
+    this.matchCase = matchCase;
+  }
+
+  public void setText(String text) {
+    this.text = text;
+  }
+}
