+++ shesmu-server/src/main/java/ca/on/oicr/gsi/shesmu/server/MetroDiagram.java
@@ -7,8 +7,9 @@ import ca.on.oicr.gsi.shesmu.compiler.description.OliveClauseRow;
 import ca.on.oicr.gsi.shesmu.compiler.description.OliveTable;
 import ca.on.oicr.gsi.shesmu.compiler.description.VariableInformation;
 import ca.on.oicr.gsi.shesmu.compiler.description.VariableInformation.Behaviour;
+import ca.on.oicr.gsi.shesmu.plugin.SourceLocation;
+import ca.on.oicr.gsi.shesmu.plugin.SourceLocation.SourceLoctionLinker;
 import ca.on.oicr.gsi.shesmu.plugin.types.Imyhat;
-import ca.on.oicr.gsi.shesmu.server.SourceLocation.SourceLoctionLinker;
 import java.awt.Canvas;
 import java.awt.Font;
 import java.time.Instant;
